PROJECT_MIGRATION/DISCORD BOT/requirements.txt
aiohttp>=3.11.18
certifi>=2025.4.26
colorama>=0.4.6
discord.py>=2.5.2
PyNaCl>=1.5.0
requests>=2.32.3
setuptools>=80.3.1
pyzipper>=0.3.6
pytz>=2025.2
aiohttp-socks>=0.10.1
python-dotenv>=1.1.0
onnxruntime>=1.18.1
matplotlib==3.10.3
arabic-reshaper==3.0.0
python-bidi==0.6.6
pillow>=10.4.0
numpy>=1.26.4
pandas>=2.0.0
google-api-python-client>=2.90.0
google-auth>=2.20.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.0.0
pymongo>=4.5.0
motor>=3.0.0
beautifulsoup4>=4.12.0
duckduckgo-search>=2.8.0
wavelink>=3.4.1
deepl>=1.18.0
openai-whisper>=20231117
gtts>=2.5.0
SpeechRecognition>=3.10.0
pydub>=0.25.1
soundfile>=0.12.1
edge-tts>=6.1.0
# Alliance Monitor Dependencies
mss>=9.0.0
opencv-python>=4.8.0
pytesseract>=0.3.10
pywin32>=306; sys_platform == 'win32'
